--- a/src/kem/kem.h
+++ b/src/kem/kem.h
@@ -357,7 +357,7 @@ OQS_API void OQS_KEM_free(OQS_KEM *kem);
 #include <oqs/kem_sike.h>
 #endif /* OQS_ENABLE_KEM_SIKE */
 #ifdef OQS_ENABLE_KEM_SIDH
-#include <oqs/kem_sike.h>
+#include <oqs/kem_sidh.h>
 #endif /* OQS_ENABLE_KEM_SIDH */
 // EDIT-WHEN-ADDING-KEM