--- a/src/kem/sike/P434/P434_compressed.c
+++ b/src/kem/sike/P434/P434_compressed.c
@@ -364,6 +364,9 @@ static const uint64_t v_3_torsion[20][2 * NWORDS64_FIELD] = {
 #define crypto_kem_keypair OQS_KEM_sike_p434_compressed_keypair
 #define crypto_kem_enc OQS_KEM_sike_p434_compressed_encaps
 #define crypto_kem_dec OQS_KEM_sike_p434_compressed_decaps
+#define OQS_KEM_sidh_keypair OQS_KEM_sidh_p434_compressed_keypair
+#define OQS_KEM_sidh_encaps OQS_KEM_sidh_p434_compressed_encaps
+#define OQS_KEM_sidh_decaps OQS_KEM_sidh_p434_compressed_decaps
 #define random_mod_order_A oqs_kem_sidh_p434_compressed_random_mod_order_A
 #define random_mod_order_B oqs_kem_sidh_p434_compressed_random_mod_order_B
 #define EphemeralKeyGeneration_A oqs_kem_sidh_p434_compressed_EphemeralKeyGeneration_A
@@ -388,4 +391,4 @@ static const uint64_t v_3_torsion[20][2 * NWORDS64_FIELD] = {
 #include "P434_compressed_dlog_tables.c"
 #include "../compression/dlog.c"
 #include "../compression/sidh_compressed.c"
-#include "../compression/sike_compressed.c"
+#include "../kem.c"