--- a/src/kem/sike/P503/P503.c
+++ b/src/kem/sike/P503/P503.c
@@ -131,6 +131,9 @@ static const unsigned int strat_Bob[MAX_Bob - 1] = {
 #define crypto_kem_keypair OQS_KEM_sike_p503_keypair
 #define crypto_kem_enc OQS_KEM_sike_p503_encaps
 #define crypto_kem_dec OQS_KEM_sike_p503_decaps
+#define OQS_KEM_sidh_keypair OQS_KEM_sidh_p503_keypair
+#define OQS_KEM_sidh_encaps OQS_KEM_sidh_p503_encaps
+#define OQS_KEM_sidh_decaps OQS_KEM_sidh_p503_decaps
 #define random_mod_order_A oqs_kem_sidh_p503_random_mod_order_A
 #define random_mod_order_B oqs_kem_sidh_p503_random_mod_order_B
 #define EphemeralKeyGeneration_A oqs_kem_sidh_p503_EphemeralKeyGeneration_A
@@ -151,4 +154,4 @@ static const unsigned int strat_Bob[MAX_Bob - 1] = {
 #include "../fpx.c"
 #include "../ec_isogeny.c"
 #include "../sidh.c"
-#include "../sike.c"
+#include "../kem.c"