--- a/src/kem/sike/fpx.c
+++ b/src/kem/sike/fpx.c
@@ -138,7 +138,7 @@ static void fp2correction(f2elm_t a) { // Modular correction, a = a in GF(p^2).
 }
 
 __inline static void mp_addfast(const digit_t *a, const digit_t *b, digit_t *c) { // Multiprecision addition, c = a+b.
-#if USE_SIKE_ASM
+#ifdef USE_SIKE_ASM
 	mp_add_asm(a, b, c);
 #else
 	mp_add(a, b, c, NWORDS_FIELD);
@@ -168,7 +168,7 @@ __inline unsigned int mp_sub(const digit_t *a, const digit_t *b, digit_t *c, con
 }
 
 __inline static void mp_subaddfast(const digit_t *a, const digit_t *b, digit_t *c) { // Multiprecision subtraction followed by addition with p*2^MAXBITS_FIELD, c = a-b+(p*2^MAXBITS_FIELD) if a-b < 0, otherwise c=a-b.
-#if USE_SIKE_ASM
+#ifdef USE_SIKE_ASM
 	mp_subaddx2_asm(a, b, c);
 #else
 	felm_t t1;
@@ -182,7 +182,7 @@ __inline static void mp_subaddfast(const digit_t *a, const digit_t *b, digit_t *
 }
 
 __inline static void mp_dblsubfast(const digit_t *a, const digit_t *b, digit_t *c) { // Multiprecision subtraction, c = c-a-b, where lng(a) = lng(b) = 2*NWORDS_FIELD.
-#if USE_SIKE_ASM
+#ifdef USE_SIKE_ASM
 	mp_dblsubx2_asm(a, b, c);
 #else
 	mp_sub(c, a, c, 2 * NWORDS_FIELD);