… CMakeLists.txt
… src/kem/sike/kem_sike.c
… src/kem/sike/kem_sike.h