Tried to solve the issue https://github.com/open-quantum-safe/liboqs/issues/737